Updates to teams work
- Move aggregations to one place and share
- Add endpoint param validation where it was missing
- Aliogn helpers folders with rest of code base
