+++ src/api/teams/controllers/add-user-to-team.js
@@ -1,7 +1,7 @@
 import Boom from '@hapi/boom'
 
 import { config } from '~/src/config/index.js'
-import { getTeam } from '~/src/api/teams/helpers/mongo/get-team.js'
+import { getTeam } from '~/src/api/teams/helpers/get-team.js'
 import { getUser } from '~/src/api/users/helpers/get-user.js'
 import { teamHasUser } from '~/src/api/teams/helpers/team-has-user.js'
 import { addUserToTeam } from '~/src/helpers/mongo/transactions/add-user-to-team.js'
