+++ src/api/teams/controllers/create-team.js
@@ -3,7 +3,7 @@ import Boom from '@hapi/boom'
 import { config } from '~/src/config/index.js'
 import { createTeamValidationSchema } from '~/src/api/teams/helpers/create-team-validation-schema.js'
 import { MongoErrors } from '~/src/helpers/mongodb-errors.js'
-import { teamNameExists } from '~/src/api/teams/helpers/mongo/team-name-exists.js'
+import { teamNameExists } from '~/src/api/teams/helpers/team-name-exists.js'
 import { gitHubTeamExists } from '~/src/api/teams/helpers/github/github-team-exists.js'
 import { createTeam } from '~/src/api/teams/helpers/aad/create-team.js'
 import { addSharedRepoAccess } from '~/src/api/teams/helpers/github/github-shared-repo-access.js'
