+++ src/api/teams/controllers/delete-team.js
@@ -1,9 +1,16 @@
-import { deleteTeam } from '~/src/helpers/mongo/transactions/delete-transactions.js'
+import Joi from 'joi'
+
 import { config } from '~/src/config/index.js'
+import { deleteTeam } from '~/src/helpers/mongo/transactions/delete-transactions.js'
 import { removeUserFromAadGroup } from '~/src/api/teams/helpers/remove-user-from-aad-group.js'
 
 const deleteTeamController = {
   options: {
+    validate: {
+      params: Joi.object({
+        teamId: Joi.string().uuid().required()
+      })
+    },
     auth: {
       strategy: 'azure-oidc',
       access: {
