+++ src/api/teams/controllers/get-team.js
@@ -1,8 +1,16 @@
 import Boom from '@hapi/boom'
 
-import { getTeam } from '~/src/api/teams/helpers/mongo/get-team.js'
+import { getTeam } from '~/src/api/teams/helpers/get-team.js'
+import Joi from 'joi'
 
 const getTeamController = {
+  options: {
+    validate: {
+      params: Joi.object({
+        teamId: Joi.string().uuid().required()
+      })
+    }
+  },
   handler: async (request, h) => {
     const team = await getTeam(request.db, request.params.teamId)
     if (!team) {
