+++ src/api/teams/helpers/aad/create-team.js
@@ -1,7 +1,7 @@
 import { config } from '~/src/config/index.js'
 import { mailNicknameFromGroupName } from '~/src/api/teams/helpers/mail-nickname-from-group-name.js'
 import { groupNameFromTeamName } from '~/src/api/teams/helpers/group-name-from-team-name.js'
-import { getTeam } from '~/src/api/teams/helpers/mongo/get-team.js'
+import { getTeam } from '~/src/api/teams/helpers/get-team.js'
 import { removeNil } from '~/src/helpers/remove-nil.js'
 
 async function createTeam(msGraph, db, dbTeam) {
