+++ src/api/teams/helpers/team-name-exists.js
