+++ src/api/teams/helpers/update-team.js
@@ -3,7 +3,7 @@ import isNull from 'lodash/isNull.js'
 
 import { mailNicknameFromGroupName } from '~/src/api/teams/helpers/mail-nickname-from-group-name.js'
 import { groupNameFromTeamName } from '~/src/api/teams/helpers/group-name-from-team-name.js'
-import { getTeam } from '~/src/api/teams/helpers/mongo/get-team.js'
+import { getTeam } from '~/src/api/teams/helpers/get-team.js'
 
 async function updateTeam(msGraph, db, teamId, updateFields) {
   const updateGroupFields = {}
